Refactor AutoTimer into the TimerService spi package.
