+++ ejb3/src/main/java/org/jboss/as/ejb3/timerservice/spi/AutoTimer.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.ejb3.timerservice;
+package org.jboss.as.ejb3.timerservice.spi;
 
 import java.lang.reflect.Method;
 import javax.ejb.ScheduleExpression;
